# docs/swagger/swagger.yaml
definitions:
  github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_response.EmptySuccessResponse:
    properties:
      message:
        type: string
    type: object
  internal_ports_http_handlers_auth.loginRequest:
    properties:
      password:
        type: string
      phone:
        maxLength: 15
        minLength: 10
        type: string
    required:
    - password
    - phone
    type: object
  internal_ports_http_handlers_auth.loginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  internal_ports_http_handlers_auth.refreshTokensRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_ports_http_handlers_auth.refreshTokensResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  internal_ports_http_handlers_auth.registerConfirmRequest:
    properties:
      name:
        minLength: 2
        type: string
      password:
        type: string
      surname:
        minLength: 2
        type: string
      token:
        type: string
    required:
    - name
    - password
    - surname
    - token
    type: object
  internal_ports_http_handlers_auth.registerConfirmResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  internal_ports_http_handlers_auth.registerStaffRequest:
    properties:
      phone:
        maxLength: 15
        minLength: 10
        type: string
      point_code:
        type: string
    required:
    - phone
    - point_code
    type: object
  internal_ports_http_handlers_form.clientFormRequest:
    properties:
      description:
        maxLength: 500
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        maxLength: 50
        minLength: 2
        type: string
      phone:
        type: string
      role:
        type: string
    required:
    - description
    - first_name
    - last_name
    - phone
    - role
    type: object
host: localhost:8080
info:
  contact:
    email: support@bagsy.com
    name: API Support
  description: API для приложения Bagsy
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bagsy API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Выполняет авторизацию пользователя по номеру телефона и паролю
      parameters:
      - description: Данные для авторизации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_ports_http_handlers_auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация
          schema:
            $ref: '#/definitions/internal_ports_http_handlers_auth.loginResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "401":
          description: Неверный логин или пароль
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Выполняет выход пользователя из системы и инвалидирует refresh
        token
      parameters:
      - description: Refresh token для выхода
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_ports_http_handlers_auth.refreshTokensRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
      summary: Выход из системы
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет токены авторизации
      parameters:
      - description: Токен авторизации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_ports_http_handlers_auth.refreshTokensRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены авторизации успешно обновлены
          schema:
            $ref: '#/definitions/internal_ports_http_handlers_auth.refreshTokensResponse'
        "400":
          description: Неверный токен
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
      summary: Обновление токенов авторизации
      tags:
      - auth
  /api/v1/auth/staff/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя (работника) привязанного к точке. Менеджер
        точки может создавать только в своей точке, менеджер сети - в любой точке
        своей сети.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_ports_http_handlers_auth.registerStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_response.EmptySuccessResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "403":
          description: Недостаточно прав для создания работника в указанной точке
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
      summary: Регистрация работника на точке
      tags:
      - auth
  /api/v1/auth/staff/register/confirm:
    post:
      consumes:
      - application/json
      description: Подтверждает регистрацию пользователя(работника)
      parameters:
      - description: Данные для подтверждения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_ports_http_handlers_auth.registerConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Регистрация успешно подтверждена
          schema:
            $ref: '#/definitions/internal_ports_http_handlers_auth.registerConfirmResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "401":
          description: Неверный или просроченный токен
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
      summary: Подтверждение регистрации работника
      tags:
      - auth
  /api/v1/forms:
    post:
      consumes:
      - application/json
      description: Создает новую заявку на сотрудничество
      parameters:
      - description: Форма для создания заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_ports_http_handlers_form.clientFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное создание заявки
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_response.EmptySuccessResponse'
        "500":
          description: ошибка
          schema:
            $ref: '#/definitions/github_com_Rasikrr_bagsy_backend_monolith_internal_ports_http_errors.ErrorResponse'
      summary: Отправка формы для сотрудничества
      tags:
      - forms
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"